+++ routes/matches.js
@@ -2,6 +2,11 @@ var express = require('express');
 var matches = express.Router();
 var queries = require('../queries');
 var config = require('../config');
+var compute = require('../compute');
+var computeMatchData = compute.computeMatchData;
+var renderMatch = compute.renderMatch;
+var redis = require('../redis').client;
+var db = require('../db');
 var matchPages = {
     index: {
         name: "Basic"
@@ -30,7 +35,7 @@ var matchPages = {
 };
 matches.get('/:match_id/:info?', function(req, res, next) {
     console.time("match page");
-    queries.prepareMatch(req.params.match_id, function(err, match) {
+    prepareMatch(req.params.match_id, function(err, match) {
         if (err) {
             return next(err);
         }
@@ -47,4 +52,45 @@ matches.get('/:match_id/:info?', function(req, res, next) {
         });
     });
 });
+
+function prepareMatch(match_id, cb) {
+    var key = "match:" + match_id;
+    redis.get(key, function(err, reply) {
+        if (!err && reply) {
+            console.log("Cache hit for match " + match_id);
+            try {
+                var match = JSON.parse(reply);
+                return cb(err, match);
+            }
+            catch (e) {
+                return cb(e);
+            }
+        }
+        else {
+            console.log("Cache miss for match " + match_id);
+            db.matches.findOne({
+                match_id: Number(match_id)
+            }, function(err, match) {
+                if (err || !match) {
+                    return cb("match not found");
+                }
+                else {
+                    queries.fillPlayerNames(match.players, function(err) {
+                        if (err) {
+                            return cb(err);
+                        }
+                        computeMatchData(match);
+                        renderMatch(match);
+                        //Add to cache if match is parsed
+                        //TODO: this prevents reparses from showing immediately
+                        if (match.parse_status === 2 && config.NODE_ENV !== "development") {
+                            redis.setex(key, 3600, JSON.stringify(match));
+                        }
+                        return cb(err, match);
+                    });
+                }
+            });
+        }
+    });
+}
 module.exports = matches;
