+++ views/player/player_activity.jade
@@ -1,9 +1,6 @@
 extend player
 
 block playercontent
-  .row
-    .col-md-12
-      include /query
   
   hr
   .row
