+++ views/player/player_matches.jade
@@ -1,0 +1,11 @@
+extend player
+
+block playercontent
+  - var clientConstants = {"game_mode": constants.game_mode,"lobby_type": constants.lobby_type,"heroes": constants.heroes,"parse_status": constants.parse_status};
+  script.
+    var constants = !{JSON.stringify(clientConstants)};
+    var matches = !{JSON.stringify(player.matches)};
+  .row
+    .col-md-12.table-responsive
+      h3 Matches
+      table.table#matches
